engine/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
anthropic>=0.40.0
supabase>=2.0.0
